+++ CHANGELOG.md
@@ -4,6 +4,8 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ## [Unreleased]
 
+## [1.0.0] - 2026-02-04
+
 ### Added
 - RDF triple support in datasets.yaml with automatic prefix expansion in datapackage.json
 - Support for rdfPrefixes at dataset level and in defaults section